camera/capture_image_cancel-manual.html

<!DOCTYPE html>
<meta charset='utf-8'>
<title>HTML Media Capture Test: user denial of captured image leading to no capture</title>
<link rel='author' title='Intel' href='http://www.intel.com'>
<link rel='help' href='http://www.w3.org/TR/html-media-capture/#security'>
<link rel='help' href='http://www.w3.org/TR/html5/forms.html#attr-input-accept'>
<meta name='flags' content='interact'>
<meta name="timeout" content="long">
<script src='../support/testharness.js'></script>
<script src='../support/testharnessreport.js'></script>

<p>Clear all camera permissions before running this test. If prompted for permission, please allow.</p>

<p>After hitting the button below, capture an image and then cancel the capturing. Please do these actions in 30 seconds.</p>

<input type='file' accept='image/*' capture>

<div id='log'></div>

<script>
  var input;

  setup(function() {
    input = document.querySelector('input[type=file]');
  }, {explicit_done: true});

  async_test(function(t) {
    input.onchange = t.unreached_func("Should not confirm the capturing");

    input.onclick = t.step_func(function() {

      // Here has a bug: after clicking the button but no more actions, it still get PASS after 30s.
      setTimeout(function() {
        assert_equals(input.files.length, 0, 'Denial of captured image leads to no capture');
        t.done();
        done();
      }, 30000);
    });
  });
</script>
